+++ pumpkin-lib/src/propagators/clausal_propagators/clausal_propagator_inline_binary.rs
@@ -125,7 +125,6 @@ impl ClausalPropagatorInterface for ClausalPropagatorInlineBinary {
             assignments.enqueue_propagated_literal(
                 asserting_literal,
                 ClauseReference::create_virtual_binary_clause_reference(second_literal).into(),
-                None,
             );
             None
         }
@@ -134,11 +133,7 @@ impl ClausalPropagatorInterface for ClausalPropagatorInlineBinary {
             let clause_reference = self
                 .add_clause_unchecked(literals, true, clause_allocator)
                 .expect("Add clause failed for some reason");
-            assignments.enqueue_propagated_literal(
-                asserting_literal,
-                clause_reference.into(),
-                None,
-            );
+            assignments.enqueue_propagated_literal(asserting_literal, clause_reference.into());
             Some(clause_reference)
         }
     }
@@ -247,7 +242,6 @@ impl ClausalPropagatorInterface for ClausalPropagatorInlineBinary {
                         assignments.enqueue_propagated_literal(
                             cached_literal,
                             watched_clause_reference.into(),
-                            None,
                         );
                         //keep the watcher
                         self.watch_lists[!true_literal][end_index] =
@@ -339,11 +333,8 @@ impl ClausalPropagatorInterface for ClausalPropagatorInlineBinary {
                 //	watched_clause[0] is assigned false -> conflict
 
                 //can propagate?
-                let conflict_info = assignments.enqueue_propagated_literal(
-                    watched_clause[0],
-                    watched_clause_reference.into(),
-                    None,
-                );
+                let conflict_info = assignments
+                    .enqueue_propagated_literal(watched_clause[0], watched_clause_reference.into());
                 if let Some(conflict_info) = conflict_info {
                     //conflict detected, stop any further propagation and report the conflict
                     //  pumpkin_assert_advanced(state_.assignments_.IsAssignedFalse(watched_clause[0]), "Sanity check.");