+++ pumpkin-lib/src/propagators/int_times.rs
@@ -1,15 +1,13 @@
-use std::collections::HashMap;
-
 use log::warn;
 
-use crate::engine::DomainEvents;
+use crate::engine::{DomainEvents, PropagationContext, PropagationContextMut, ReadDomains};
 use crate::{
-    basic_types::{variables::IntVar, Predicate, PropagationStatusCP, PropositionalConjunction},
+    basic_types::{variables::IntVar, PropagationStatusCP},
+    conjunction,
     engine::{
-        CPPropagatorConstructor, ConstraintProgrammingPropagator, Delta, DomainChange, LocalId,
-        PropagationContext, PropagatorConstructorContext, PropagatorVariable,
+        CPPropagatorConstructor, ConstraintProgrammingPropagator, LocalId,
+        PropagatorConstructorContext, PropagatorVariable,
     },
-    predicate,
 };
 
 /// A bounds-consistent propagator for maintaining the constraint `a * b = c`. The propagator
@@ -18,10 +16,6 @@ pub struct IntTimesProp<VA, VB, VC> {
     a: PropagatorVariable<VA>,
     b: PropagatorVariable<VB>,
     c: PropagatorVariable<VC>,
-
-    propagations_a: HashMap<DomainChange, [Predicate; 2]>,
-    propagations_b: HashMap<DomainChange, [Predicate; 2]>,
-    propagations_c: HashMap<DomainChange, [Predicate; 2]>,
 }
 
 pub struct IntTimes<VA, VB, VC> {
@@ -47,10 +41,6 @@ where
             a: context.register(self.a, DomainEvents::ANY_INT, ID_A),
             b: context.register(self.b, DomainEvents::ANY_INT, ID_B),
             c: context.register(self.c, DomainEvents::ANY_INT, ID_C),
-
-            propagations_a: Default::default(),
-            propagations_b: Default::default(),
-            propagations_c: Default::default(),
         }
     }
 }
@@ -61,51 +51,11 @@ where
     VB: IntVar,
     VC: IntVar,
 {
-    fn propagate(&mut self, context: &mut PropagationContext) -> PropagationStatusCP {
-        perform_propagation(
-            context,
-            &self.a,
-            &self.b,
-            &self.c,
-            &mut self.propagations_a,
-            &mut self.propagations_b,
-            &mut self.propagations_c,
-        )
+    fn propagate(&mut self, context: &mut PropagationContextMut) -> PropagationStatusCP {
+        perform_propagation(context, &self.a, &self.b, &self.c)
     }
 
-    fn synchronise(&mut self, context: &PropagationContext) {
-        fn cleanup<Var: IntVar>(
-            context: &PropagationContext,
-            var: &PropagatorVariable<Var>,
-            propagations: &mut HashMap<DomainChange, [Predicate; 2]>,
-        ) {
-            propagations.retain(|change, _reason| match change {
-                DomainChange::Removal(value) => !context.contains(var, *value),
-                DomainChange::LowerBound(bound) => context.lower_bound(var) >= *bound,
-                DomainChange::UpperBound(bound) => context.upper_bound(var) <= *bound,
-                _ => unreachable!(),
-            });
-        }
-
-        cleanup(context, &self.a, &mut self.propagations_a);
-        cleanup(context, &self.b, &mut self.propagations_b);
-        cleanup(context, &self.c, &mut self.propagations_c);
-    }
-
-    fn get_reason_for_propagation(
-        &mut self,
-        _context: &PropagationContext,
-        delta: Delta,
-    ) -> PropositionalConjunction {
-        let reason = match delta.affected_local_id() {
-            ID_A => self.propagations_a[&self.a.unpack(delta)],
-            ID_B => self.propagations_b[&self.b.unpack(delta)],
-            ID_C => self.propagations_c[&self.c.unpack(delta)],
-            _ => unreachable!(),
-        };
-
-        PropositionalConjunction::from(reason.to_vec())
-    }
+    fn synchronise(&mut self, _context: &PropagationContext) {}
 
     fn priority(&self) -> u32 {
         0
@@ -115,7 +65,7 @@ where
         "IntTimes"
     }
 
-    fn initialise_at_root(&mut self, context: &mut PropagationContext) -> PropagationStatusCP {
+    fn initialise_at_root(&mut self, context: &mut PropagationContextMut) -> PropagationStatusCP {
         if context.lower_bound(&self.a).is_negative() || context.lower_bound(&self.b).is_negative()
         {
             warn!("IntTimes does not support variables with negative bounds, but it is instantiated with them. The behavior will likely not be correct.");
@@ -126,32 +76,17 @@ where
 
     fn debug_propagate_from_scratch(
         &self,
-        context: &mut PropagationContext,
+        context: &mut PropagationContextMut,
     ) -> PropagationStatusCP {
-        let mut propagations_a = HashMap::new();
-        let mut propagations_b = HashMap::new();
-        let mut propagations_c = HashMap::new();
-
-        perform_propagation(
-            context,
-            &self.a,
-            &self.b,
-            &self.c,
-            &mut propagations_a,
-            &mut propagations_b,
-            &mut propagations_c,
-        )
+        perform_propagation(context, &self.a, &self.b, &self.c)
     }
 }
 
 fn perform_propagation<VA: IntVar, VB: IntVar, VC: IntVar>(
-    context: &mut PropagationContext,
+    context: &mut PropagationContextMut,
     a: &PropagatorVariable<VA>,
     b: &PropagatorVariable<VB>,
     c: &PropagatorVariable<VC>,
-    propagations_a: &mut HashMap<DomainChange, [Predicate; 2]>,
-    propagations_b: &mut HashMap<DomainChange, [Predicate; 2]>,
-    propagations_c: &mut HashMap<DomainChange, [Predicate; 2]>,
 ) -> PropagationStatusCP {
     let a_min = context.lower_bound(a);
     let a_max = context.upper_bound(a);
@@ -175,30 +110,18 @@ fn perform_propagation<VA: IntVar, VB: IntVar, VC: IntVar>(
     let new_min_c = a_min * b_min;
 
     if context.upper_bound(c) > new_max_c {
-        propagations_c.insert(
-            DomainChange::UpperBound(new_max_c),
-            [predicate![a <= a_max], predicate![b <= b_max]],
-        );
-        context.set_upper_bound(c, new_max_c)?;
+        context.set_upper_bound(c, new_max_c, conjunction!([a <= a_max] & [b <= b_max]))?;
     }
 
     if context.lower_bound(c) < new_min_c {
-        propagations_c.insert(
-            DomainChange::LowerBound(new_min_c),
-            [predicate![a >= a_min], predicate![b >= b_min]],
-        );
-        context.set_lower_bound(c, new_min_c)?;
+        context.set_lower_bound(c, new_min_c, conjunction!([a >= a_min] & [b >= b_min]))?;
     }
 
     // a >= ceil(c.min / b.max)
     if b_max >= 1 {
         let bound = div_ceil_pos(c_min, b_max);
         if context.lower_bound(a) < bound {
-            propagations_a.insert(
-                DomainChange::LowerBound(bound),
-                [predicate![c >= c_min], predicate![b <= b_max]],
-            );
-            context.set_lower_bound(a, bound)?;
+            context.set_lower_bound(a, bound, conjunction!([c >= c_min] & [b <= b_max]))?;
         }
     }
 
@@ -206,11 +129,7 @@ fn perform_propagation<VA: IntVar, VB: IntVar, VC: IntVar>(
     if b_min >= 1 {
         let bound = c_max / b_min;
         if context.upper_bound(a) > bound {
-            propagations_a.insert(
-                DomainChange::UpperBound(bound),
-                [predicate![c <= c_max], predicate![b >= b_min]],
-            );
-            context.set_upper_bound(a, bound)?;
+            context.set_upper_bound(a, bound, conjunction!([c <= c_max] & [b >= b_min]))?;
         }
     }
 
@@ -219,11 +138,7 @@ fn perform_propagation<VA: IntVar, VB: IntVar, VC: IntVar>(
         let bound = div_ceil_pos(c_min, a_max);
 
         if context.lower_bound(b) < bound {
-            propagations_b.insert(
-                DomainChange::LowerBound(bound),
-                [predicate![c >= c_min], predicate![a <= a_max]],
-            );
-            context.set_lower_bound(b, bound)?;
+            context.set_lower_bound(b, bound, conjunction!([c >= c_min] & [a <= a_max]))?;
         }
     }
 
@@ -231,11 +146,7 @@ fn perform_propagation<VA: IntVar, VB: IntVar, VC: IntVar>(
     if a_min >= 1 {
         let bound = c_max / a_min;
         if context.upper_bound(b) > bound {
-            propagations_b.insert(
-                DomainChange::UpperBound(bound),
-                [predicate![c <= c_max], predicate![a >= a_min]],
-            );
-            context.set_upper_bound(b, bound)?;
+            context.set_upper_bound(b, bound, conjunction!([c <= c_max] & [a >= a_min]))?;
         }
     }
     Ok(())
@@ -251,7 +162,7 @@ fn div_ceil_pos(numerator: i32, denominator: i32) -> i32 {
 
 #[cfg(test)]
 mod tests {
-    use crate::{conjunction, engine::test_helper::TestSolver};
+    use crate::{conjunction, engine::test_helper::TestSolver, predicate};
 
     use super::*;
 
@@ -275,17 +186,11 @@ mod tests {
         assert_eq!(0, solver.lower_bound(c));
         assert_eq!(12, solver.upper_bound(c));
 
-        let reason_lb = solver.get_reason(
-            &mut propagator,
-            Delta::new(ID_C, DomainChange::LowerBound(0)),
-        );
-        assert_eq!(conjunction!([a >= 1] & [b >= 0]), reason_lb);
+        let reason_lb = solver.get_reason_int(predicate![c >= 0]);
+        assert_eq!(conjunction!([a >= 1] & [b >= 0]), *reason_lb);
 
-        let reason_ub = solver.get_reason(
-            &mut propagator,
-            Delta::new(ID_C, DomainChange::UpperBound(12)),
-        );
-        assert_eq!(conjunction!([a <= 3] & [b <= 4]), reason_ub);
+        let reason_ub = solver.get_reason_int(predicate![c <= 12]);
+        assert_eq!(conjunction!([a <= 3] & [b <= 4]), *reason_ub);
     }
 
     #[test]
@@ -308,17 +213,11 @@ mod tests {
         assert_eq!(2, solver.lower_bound(c));
         assert_eq!(12, solver.upper_bound(c));
 
-        let reason_lb = solver.get_reason(
-            &mut propagator,
-            Delta::new(ID_B, DomainChange::LowerBound(1)),
-        );
-        assert_eq!(conjunction!([a <= 3] & [c >= 2]), reason_lb);
+        let reason_lb = solver.get_reason_int(predicate![b >= 1]);
+        assert_eq!(conjunction!([a <= 3] & [c >= 2]), *reason_lb);
 
-        let reason_ub = solver.get_reason(
-            &mut propagator,
-            Delta::new(ID_B, DomainChange::UpperBound(6)),
-        );
-        assert_eq!(conjunction!([a >= 2] & [c <= 12]), reason_ub);
+        let reason_ub = solver.get_reason_int(predicate![b <= 6]);
+        assert_eq!(conjunction!([a >= 2] & [c <= 12]), *reason_ub);
     }
 
     #[test]
@@ -341,16 +240,10 @@ mod tests {
         assert_eq!(3, solver.lower_bound(c));
         assert_eq!(12, solver.upper_bound(c));
 
-        let reason_lb = solver.get_reason(
-            &mut propagator,
-            Delta::new(ID_A, DomainChange::LowerBound(1)),
-        );
-        assert_eq!(conjunction!([b <= 6] & [c >= 2]), reason_lb);
+        let reason_lb = solver.get_reason_int(predicate![a >= 1]);
+        assert_eq!(conjunction!([b <= 6] & [c >= 2]), *reason_lb);
 
-        let reason_ub = solver.get_reason(
-            &mut propagator,
-            Delta::new(ID_A, DomainChange::UpperBound(4)),
-        );
-        assert_eq!(conjunction!([b >= 3] & [c <= 12]), reason_ub);
+        let reason_ub = solver.get_reason_int(predicate![a <= 4]);
+        assert_eq!(conjunction!([b >= 3] & [c <= 12]), *reason_ub);
     }
 }